backend/requirements.txt
# Core Flask
Flask==2.3.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.5

# Database
Flask-SQLAlchemy==3.0.3
SQLAlchemy==2.0.19
psycopg2-binary==2.9.7  # PostgreSQL adapter

# Authentication & Security
PyJWT==2.8.0
Flask-JWT-Extended==4.5.2
Werkzeug==2.3.7  # Includes password hashing
bcrypt==4.0.1  # Alternative password hashing

# Environment & Configuration
python-dotenv==1.0.0

# API Development
Flask-RESTful==0.3.10  # Optional for REST API structure
marshmallow==3.19.0  # Optional for serialization

# File Uploads
Pillow==10.0.1  # For image processing

# Development & Testing
pytest==7.4.0
Flask-Testing==0.8.1
requests==2.31.0

# Production
gunicorn==21.2.0